feat(nx-dev): update youtube cmp to allow caption
