--- nx-dev/ui-markdoc/src/lib/tags/youtube.schema.ts
@@ -1,19 +1,0 @@
-import { Schema } from '@markdoc/markdoc';
-
-export const youtube: Schema = {
-  render: 'YouTube',
-  attributes: {
-    src: {
-      type: 'String',
-      required: true,
-    },
-    title: {
-      type: 'String',
-      required: true,
-    },
-    width: {
-      type: 'String',
-      default: '50%',
-    },
-  },
-};
